requirements.txt
dimod
matplotlib
dwave-samplers
flask